ebsi_sim/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bitarray==3.8.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
ckzg==2.1.7
click==8.3.2
contextvars==2.4
cryptography==46.0.7
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==6.0.0b1
fastapi[standard]==0.136.0
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
immutables==0.21
iniconfig==2.3.0
Jinja2==3.1.6
jsonpath-ng==1.8.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwcrypto==1.5.7
lxml==6.0.4
markdown2==2.5.5
MarkupSafe==3.0.3
multidict==6.7.1
packaging==26.1
parsimonious==0.10.0
pathlib==1.0.1
pdoc==16.0.0
pluggy==1.6.0
ply==3.11
propcache==0.4.1
psycopg2-binary==2.9.11
py-ecc==8.0.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
python-docx==1.2.0
python-dotenv==1.2.2
pyunormalize==17.0.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rlp==4.1.0
rpds-py==0.30.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
toolz==1.1.0
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
web3==7.15.0
websockets==15.0.1
yarl==1.23.0
